# docs/swagger.yaml
definitions:
  structs.Avatar:
    properties:
      body:
        type: string
      face:
        type: string
      head:
        type: string
    type: object
  structs.Error:
    properties:
      error:
        type: string
    type: object
  structs.JWTResponse:
    properties:
      success:
        type: boolean
    type: object
  structs.Me:
    properties:
      session_expire_at:
        type: integer
      user_id:
        type: string
    type: object
  structs.Profile:
    properties:
      avatar:
        $ref: '#/definitions/structs.Avatar'
      currency:
        type: integer
      experience:
        type: number
      id:
        type: integer
      initial_login_date:
        type: string
      is_bedrock:
        type: boolean
      last_login_date:
        type: string
      name:
        type: string
      total_build_blocks:
        type: integer
      total_destroy_blocks:
        type: integer
      total_mob_kills:
        type: integer
      total_play_time:
        type: integer
      uuid:
        type: string
      xuid:
        type: string
    type: object
  structs.Server:
    properties:
      address:
        type: string
      description:
        type: string
      port:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/discord/{uuid}:
    get:
      description: fetch UUID from provided Discord ID
      parameters:
      - description: who to retrieve
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Error'
      summary: fetch UUID from provided Discord ID
      tags:
      - discord
  /auth:
    get:
      consumes:
      - application/json
      description: Login with Discord
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Location:
              type: string
      summary: Login with Discord
      tags:
      - login
  /auth/callback:
    get:
      consumes:
      - application/json
      description: callback endpoint for Discord login
      parameters:
      - description: Bearer token
        in: query
        name: code
        required: true
        type: string
      - description: Random state for validating request
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Auth-Token:
              type: string
          schema:
            items:
              $ref: '#/definitions/structs.JWTResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Error'
      summary: callback endpoint for Discord login
      tags:
      - login
  /avatar/{part}/bedrock/{xuid}:
    get:
      description: Get the specified part of player's skin image
      parameters:
      - description: which part to retrieve
        in: path
        name: part
        required: true
        type: string
      - description: XUID of target Bedrock player
        in: path
        name: xuid
        required: true
        type: string
      produces:
      - image/png
      responses:
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Error'
      summary: Get player's skin image
      tags:
      - avatar
  /me:
    get:
      consumes:
      - application/json
      description: retrieve information of authenticated user
      parameters:
      - description: JSON Web Token
        in: header
        name: X-Auth-Token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Me'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Error'
      summary: retrieve information of authenticated user
      tags:
      - login
  /profiles:
    get:
      consumes:
      - application/json
      description: Get a list of profiles with optional filtering and sorting, etc.
      parameters:
      - default: ""
        description: Filter criteria
        example: 550e8400-e29b-41d4-a716-446655440000
        in: query
        name: filter
        type: string
      - default: asc
        description: Sort order
        example: desc
        in: query
        name: sort
        type: string
      - default: 0
        description: Offset for pagination
        example: 0
        in: query
        name: offset
        type: integer
      - default: 50
        description: Limit for pagination
        example: 10
        in: query
        name: limit
        type: integer
      - description: Order by field
        example: play_time
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Profile'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Error'
      summary: Get profiles
      tags:
      - profiles
  /profiles/{uuid}:
    get:
      consumes:
      - application/json
      description: Get a profile by UUID
      parameters:
      - description: UUID of target profile
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Profile'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Error'
      summary: Get profile
      tags:
      - profiles
  /servers:
    get:
      consumes:
      - application/json
      description: Get servers registered to uzsk-api
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Server'
            type: array
      summary: Get servers
      tags:
      - servers
  /servers/{name}:
    get:
      consumes:
      - application/json
      description: Get servers registered to uzsk-api
      parameters:
      - description: Name of target server
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Server'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Error'
      summary: Get server
      tags:
      - servers
swagger: "2.0"